refactor(MainExtensions): main process extensions separated out. Less code to be run incidentally in renderers via extending
