+++ app/extensions/safe/main-process/index.ts
@@ -1,0 +1,3 @@
+export { onReceiveUrl } from './onReceiveUrl';
+export { preAppLoad } from './preAppLoad';
+export { onAppReady } from './onAppReady';
